mon: move "df" dump code from PGMonitor to PGMap
...so that ceph-mgr can use it too.
