+++ src/mon/PGMonitor.h
@@ -128,14 +128,6 @@ private:
 			  int threshold,
 			  vector<string>& args) const;
 
-  static void dump_object_stat_sum(TextTable &tbl, Formatter *f,
-				   object_stat_sum_t &sum,
-				   uint64_t avail,
-				   float raw_used_rate,
-				   bool verbose, const pg_pool_t *pool);
-
-  int64_t get_rule_avail(OSDMap& osdmap, int ruleno) const;
-
 public:
   PGMonitor(Monitor *mn, Paxos *p, const string& service_name)
     : PaxosService(mn, p, service_name),
@@ -171,9 +163,6 @@ public:
   void tick();  // check state, take actions
 
   void check_osd_map(epoch_t epoch);
-
-  void dump_pool_stats(stringstream &ss, Formatter *f, bool verbose);
-  void dump_fs_stats(stringstream &ss, Formatter *f, bool verbose) const;
 
   void dump_info(Formatter *f) const;
